AI/requirements.txt
# AI Company Holding OS — Dependencies

# LangGraph / LangChain（脳）
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# LINE Bot SDK
line-bot-sdk>=3.0.0

# Web / API
flask>=3.0.0
httpx>=0.27.0
requests>=2.32.0

# Scheduler
apscheduler>=3.10.0

# Database（状態永続化）
supabase>=2.0.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0  # リトライ
loguru>=0.7.0    # ロギング

# Dev
pytest>=8.0.0
pytest-asyncio>=0.23.0
